fix(mac): show semester boundaries and holidays in the calendar
The Mac calendar read `DataCache.loadCalendarEvents()` and nothing else,
but the academic rows are never cached — `CalendarViewModel` rebuilds
them from the in-memory calendar on every load, because a holiday name
is locale-dependent and the app language can change under us. So the
Mac had no semester boundaries and no holidays at all, and the two
switch statements that colour and label them were unreachable.

Second, narrower bug behind the same symptom: `DayCell.dedupedSources`
filters through an `order` array that listed only four of the six
sources, so a holiday would have got no dot in the month grid even once
the rows arrived.

`academicEvents` moves off `CalendarViewModel` — excluded from the macOS
build, since it pulls in EventKit that the Mac surface stays out of to
avoid the Calendar TCC entitlement — onto `AcademicCalendar` itself, so
both surfaces build their rows from one place and cannot drift.

No 'still remind me' toggle on the Mac, as before: macOS delivers no
class reminders, so there is nothing for it to switch back on. That was
already the intent — the comments saying so were just describing rows
that never actually appeared.
